Increase maximum number of open file descriptors for dnsmasq
dnsmasq has problems releasing it's file descriptors when it reaches
the maximum number of open file descriptors, this is caused by the
OpenShift Extended Comformance Tests followed by the Node Vertical Test.

This patch mitigates the problem by increasing the maximim number of
open file descriptors so the node can pass the tests.

NOTE:
The real fix should be done in dnsmasq.

Fixes: rhbz#1608571